+++ src/main/java/bdv/tools/boundingbox/TransformedBoxOverlay.java
@@ -46,7 +46,7 @@ import net.imglib2.Interval;
 import net.imglib2.RealInterval;
 import net.imglib2.realtransform.AffineTransform3D;
 import net.imglib2.ui.OverlayRenderer;
-import net.imglib2.ui.TransformListener;
+import bdv.viewer.TransformListener;
 
 import org.scijava.listeners.ChangeListener;
 import org.scijava.listeners.ListenableVar;
