+++ src/main/java/bdv/viewer/render/MultiResolutionRenderer.java
@@ -37,7 +37,6 @@ import net.imglib2.cache.iotiming.CacheIoTiming;
 import net.imglib2.realtransform.AffineTransform3D;
 import net.imglib2.type.numeric.ARGBType;
 import net.imglib2.ui.PainterThread;
-import net.imglib2.ui.RenderTarget;
 import net.imglib2.util.Intervals;
 
 import bdv.cache.CacheControl;
